use sitemap package slightly more efficiently
