+++ api/src/catalog/catalog.utils.ts
@@ -1,6 +1,8 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
 
+import { Readable } from 'node:stream';
+import { SitemapStream, streamToPromise } from 'sitemap';
 import {
   getSdk,
   type CatalogBySeasonQuery,
@@ -9,6 +11,9 @@ import {
 import { STATIC_FILE_DIR, graphqlClient } from '../config.js';
 import winston from '../logging/winston.js';
 
+interface SitemapLising {
+  crn: string;
+}
 /**
  * This is the legacy "flat" data format we used. This shape seems to be easier
  * to work with, and for the purpose of API compatibility we "massage" the GQL
@@ -119,6 +124,37 @@ async function fetchData(
   }
 }
 
+function transformListingToSitemapListing(
+  listing: CatalogBySeasonQuery['listings'][number],
+): SitemapLising {
+  return {
+    crn: String(listing.crn),
+  };
+}
+
+async function generateSeasonSitemap(
+  seasonCode: string,
+  courses: SitemapLising[],
+): Promise<void> {
+  const sitemapDir = path.join(STATIC_FILE_DIR, 'sitemaps');
+  await fs.mkdir(sitemapDir, { recursive: true });
+
+  const links = courses.map((course) => ({
+    url: `/catalog?course-modal=${seasonCode}-${course.crn}`,
+    priority: 0.8,
+  }));
+
+  const stream = new SitemapStream({ hostname: 'https://coursetable.com' });
+  const xml = await streamToPromise(Readable.from(links).pipe(stream)).then(
+    (data: Buffer) => data.toString(),
+  );
+
+  const sitemapPath = path.join(sitemapDir, `sitemap_${seasonCode}.xml`);
+  await fs.writeFile(sitemapPath, xml, 'utf-8');
+
+  winston.info(`Sitemap generated for ${seasonCode} at ${sitemapPath}`);
+}
+
 export async function fetchCatalog(overwrite: boolean) {
   let seasons: string[] = [];
   try {
@@ -158,4 +194,15 @@ export async function fetchCatalog(overwrite: boolean) {
     ),
   );
   await Promise.all(processSeasons);
+
+  for (const seasonCode of seasons) {
+    try {
+      const sdk = getSdk(graphqlClient);
+      const data = await sdk.catalogBySeason({ season: seasonCode });
+      const courses = data.listings.map(transformListingToSitemapListing);
+      await generateSeasonSitemap(seasonCode, courses);
+    } catch (err) {
+      winston.error(`Error generating sitemap for ${seasonCode}: ${err}`);
+    }
+  }
 }
